refactor: move getAngle to own file
