+++ src/maskAnalysis/utils/__tests__/getMbrFromPoints.test.ts
@@ -1,4 +1,4 @@
-import { getAngle, getMbrFromPoints } from '../getMbrFromPoints';
+import { getMbrFromPoints } from '../getMbrFromPoints';
 
 test.each([
   [
@@ -176,36 +176,4 @@ test.each([
 ])('getMbrFromPoints (%s)', (_, inputPoints, expectedPoints) => {
   const result = getMbrFromPoints(inputPoints).corners;
   expect(result).toBeDeepCloseTo(expectedPoints);
-});
-
-test.each([
-  [
-    '-60 degrees',
-    { column: 0, row: 0 },
-    { column: 1, row: -Math.sqrt(3) },
-    -Math.PI / 3,
-  ],
-  ['-45 degrees', { column: 0, row: 0 }, { column: 1, row: -1 }, -Math.PI / 4],
-  [
-    '-30 degrees',
-    { column: 0, row: 0 },
-    { column: Math.sqrt(3), row: -1 },
-    -Math.PI / 6,
-  ],
-  [
-    '60 degrees',
-    { column: 0, row: 0 },
-    { column: 1, row: Math.sqrt(3) },
-    Math.PI / 3,
-  ],
-  ['45 degrees', { column: 0, row: 0 }, { column: 1, row: 1 }, Math.PI / 4],
-  [
-    '30 degrees',
-    { column: 0, row: 0 },
-    { column: Math.sqrt(3), row: 1 },
-    Math.PI / 6,
-  ],
-])('getAngle (%s)', (_, point1, point2, expectedAngle) => {
-  const result = getAngle(point1, point2);
-  expect(result).toBeCloseTo(expectedAngle);
 });
